Split others in Summary by status
The "others" group in the Summary was only showing things not done (either todo
or wip), causing confusion as it diverges from number on the Plan.

A separate problem remains
(https://github.com/vector-im/feature-dashboard/issues/48) that we don't include
the other group in the percent complete, which may also be confusing.

Fixes https://github.com/vector-im/feature-dashboard/issues/47